### 2019-12-09 a622410
adds dois to references
