+++ paper.bib
@@ -6,14 +6,15 @@
   number={4489},
   pages={1390--1396},
   year={1981},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.7466396},
 }
 
 @misc{Weir2009,
   title={Approaches to learning: a guide for teachers},
   author={Weir, Jennifer},
   year={2009},
-  publisher={ROUTLEDGE JOURNALS, TAYLOR \& FRANCIS LTD 4 PARK SQUARE, MILTON PARK~…}
+  publisher={ROUTLEDGE JOURNALS, TAYLOR \& FRANCIS LTD 4 PARK SQUARE, MILTON PARK~…},
 }
 
 @article{Freeman2014,
@@ -24,7 +25,8 @@
   number={23},
   pages={8410--8415},
   year={2014},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={https://doi.org/10.1073/pnas.1319030111}
 }
 
 @misc{axelrodproject,
@@ -44,7 +46,8 @@
   number={3},
   pages={259--268},
   year={2004},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={https://doi.org/10.3200/JECE.35.3.259-268}
 }
 
 @article{Knight2015,
@@ -54,5 +57,6 @@
   volume={14},
   number={1},
   pages={28--38},
-  year={2015}
+  year={2015},
+  doi={https://doi.org/10.21100/msor.v14i1.254}
 }
