### 2019-12-09 #73 @Nikoleta-v3
Potential reviewers include:

- `marcharper`
- `eric-s-s`
- `gaffney2010`

These individuals are collaborators of `Axelrod-Python` which is the main software used in the tutorials. Please note that `marcharper` is also a mentor and a collaborator of mine and there is possible a conflict of interest.
