### 2020-01-20 #73 @Nikoleta-v3
Hello @labarba and @kyleniemeyer ,

I understand that you must be extremely busy, but I was wondering when an editor will be assigned to our paper?
