### 2020-02-09 #73 @labarba
hey, @Nikoleta-v3 — I do apologize for letting this submission linger in the queue for so long. It's all my fault, for taking on some big service commitments and giving lower priority to my own initiatives. Here's one big reason: I'm JupyterCon 2020 General Chair! We [just announced](https://twitter.com/JupyterCon/status/1225901511437410304?s=20). Anyway, no excuse.

👋 @racheltho — I wonder if you could take this as your first JOSE assignment as editor? I would be of course ready to help at every step of the way
Submission title: **"Game Theory and Python: An educational tutorial to game theory and repeated games using Python"**
