### 2020-02-21 #73 @labarba
After assigning the reviewers, you will do `@whedon start review`, which will create a new REVIEW issue, with the reviewer checklists, and will give the reviewers the necessary permissions to check off items in their list.
