### 2020-02-21 #73 @labarba
Now that you are editor, you have the "powers" to assign/add reviewers, and start the review.
