### 2020-02-21 #73 @labarba
You can contact potential reviewers via email, Twitter, or mention them in this issue to invite them. Once they agree, you will run `@whedon assign <user> as reviewer` for the first one, and `@whedon add <user> as reviewer` for the second one (we usually want two reviewers).

More help in our Documentation's [Editorial Guide](https://openjournals.readthedocs.io/en/jose/editing.html)
