### 2020-02-22 #73 @racheltho
Hi @sconde, you signed up to review for JOSE. Would you like to contribute a review for this submission? The title is: Game Theory and Python: An educational tutorial to game theory and repeated games using Python. Let me know if you're available and willing. Thanks!
