### 2020-02-24 #73 @labarba
Great going with your first editing for JOSE, @racheltho 🚀

A pro tip: you should find a second reviewer and run `@whedon add <user> as reviewer` here, _before_ running `@whedon start review`. That way, the review issue will be started with an individual checklist for each reviewer. Godspeed!
