### 2020-02-24 #73 @racheltho
Hi @gvdr, you signed up to review for JOSE. Would you be able to be a reviewer for this submission? The title is: Game Theory and Python: An educational tutorial to game theory and repeated games using Python. Let me know if you're available and willing. Thanks!
