### 2020-02-24 #73 @gvdr
Yes, will do!
----
Giulio Valentino Dalla Riva
Lecturer in Data Science
School of Mathematics and Statistics
University of Canterbury
Ōtautahi, Aotearoa -- Christchurch, New Zealand

----
Securely sent with Tutanota. Get your own encrypted, ad-free mailbox:
https://tutanota.com

25 Feb 2020, 11:44 by notifications@github.com:

>
> Hi > @gvdr <https://github.com/gvdr>> , you signed up to review for JOSE. Would you be able to be a reviewer for this submission? The title is: Game Theory and Python: An educational tutorial to game theory and repeated games using Python. Let me know if you're available and willing. Thanks!
>
>
