### 2020-02-25 #78 @whedon
[REVIEW]: Game Theory and Python: An educational tutorial to game theory and repeated games using Python
**Submitting author:** @Nikoleta-v3 (<a href="http://orcid.org/0000-0002-2943-3622">Nikoleta E. Glynatsi</a>)
**Repository:** <a href="https://github.com/Nikoleta-v3/Game-Theory-and-Python" target ="_blank">https://github.com/Nikoleta-v3/Game-Theory-and-Python</a>
**Version:** v1.0.0
