### 2020-02-25 #73 @arfon
> @openjournals/dev We have a delay here in `@whedon` starting the review. Can you check for us what could be the snag?

Apologies for the minor hiccup. Things should be working normally now.
