### 2020-03-06 Nikoleta-v3/Game-Theory-and-Python#22 @sconde
Statement of Need
"The authors should clearly state what problems the software is designed to solve and who the target audience is."

In the *Introduction*, you do a great job describing the problem. However, It is not clear who the intended target audience is from text.
