### 2020-03-06 Nikoleta-v3/Game-Theory-and-Python#23 @sconde
Version: no release
>  Version: Does the release version given match the repository release (v.0.0.1)?

Please create a release and update the version accordingly
