### 2020-03-06 Nikoleta-v3/Game-Theory-and-Python#24 @sconde
hard coded prefix
https://github.com/Nikoleta-v3/Game-Theory-and-Python/blob/a62241001902f2bb39765750dbcc89596167a983/environment.yml#L13

This is very specific to your machine. Please update for easy installation for other users. Or note properly in README to be updated.
