### 2020-03-06 #78 @sconde
https://github.com/Nikoleta-v3/Game-Theory-and-Python/issues/24
