### 2020-03-06 Nikoleta-v3/Game-Theory-and-Python#25 @sconde
Usage: mini presentation
"For each notebook it is advised that the instructor gives a mini presentation to the topic followed by them typing out/running the material while the participants follow in their own machines."

Have you presented this material at a workshop before? If so, it would be beneficial to provide a link to the mini presentation for other instructors to adapt.
