### 2020-03-09 #78 @Nikoleta-v3
Thank you for the issues @sconde ! 😄  I shall address them asap.
