### 2020-03-09 8a0573e
remove hard code prefix
Closes #24

After a bit of investigation it appears that conda does not care
about the prefix line when creating an environment from file. Thus,
I removed the hard code path and tested that the environment it's
still being created without issues
