+++ environment.yml
@@ -10,4 +10,3 @@ dependencies:
     - docopt
     - nbval==0.9.0
     - fsspec>=0.4.3
-prefix: /home/nikoleta/anaconda3/envs/game-python
