### 2020-03-10 Nikoleta-v3/Game-Theory-and-Python#23 @Nikoleta-v3
Thank you for the issue. I have now made a release of [v0.0.1](https://github.com/Nikoleta-v3/Game-Theory-and-Python/releases).
