### 2020-03-10 5eaa58d
rephase presentation sentence.
Closes #25
