### 2020-03-10 09fe31e
rephase presentation sentence.
Closes #25
