### 2020-03-12 74dccc7
Jose review (#26)
* remove hard code prefix

Closes #24

After a bit of investigation it appears that conda does not care
about the prefix line when creating an environment from file. Thus,
I removed the hard code path and tested that the environment it's
still being created without issues

* rephase presentation sentence.

Closes #25

* clearly states the targe audience of the tutorial in the statement of need

Closes #22

* replaces in with using

Co-Authored-By: Vince Knight <vincent.knight@gmail.com>

* bullet point target audience

Co-Authored-By: Vince Knight <vincent.knight@gmail.com>


[…]
