### 2020-03-12 #78 @Nikoleta-v3
@sconde I believe I have now addressed all the issues that you raised 👍

This was done in https://github.com/Nikoleta-v3/Game-Theory-and-Python/pull/26, and I have also made a new release (https://github.com/Nikoleta-v3/Game-Theory-and-Python/releases).
