### 2020-04-21 Nikoleta-v3/Game-Theory-and-Python#27 @sconde
JOSE Paper: tell the "story"
> Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?

The paper does not clearly tell the narrate the story.
