### 2020-04-21 #78 @sconde
> * Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?

https://github.com/Nikoleta-v3/Game-Theory-and-Python/issues/27
