### 2020-04-21 Nikoleta-v3/Game-Theory-and-Python#28 @sconde
JOSE: matches and tournaments
For independent learner, it would be very helpful to annotate the notebooks with verbose comments and/or explanations for the output of the commands.

Without some comments, the modules may not be self contained for self learner. For example, a small explanation for the result of the match **Cooperator vs Random**

```
[(C, D), (C, D), (C, C), (C, C), (C, D)]
```
