### 2020-04-21 #78 @sconde
https://github.com/Nikoleta-v3/Game-Theory-and-Python/issues/28
