### 2020-04-21 Nikoleta-v3/Game-Theory-and-Python#27 @sconde
> Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?

A little description under **Usage** subsection to discuss the above requirement is encouraged.
