### 2020-04-21 #78 @Nikoleta-v3
Thank you @sconde . I will address them asap.
