### 2020-04-25 476d4ff
minor copy edits
