### 2020-04-25 b9d0ed3
minor copy edits
