### 2020-04-25 ae4c522
adds explanations for the output of notebooks 2 and 3.
Closes #28
