### 2020-04-25 caabc4e
adds sentence about a story
Closes #27
