+++ paper.md
@@ -30,7 +30,10 @@ interactive tutorial.
 
 # Statement of Need
 
-This tutorial is aimed at two groups of individuals:
+We are mathematicians and Python programmers with interests in game theory and
+pedagogy. This tutorial was created as a resource for computer workshops and for
+introducing mathematicians to programming. As a result, this tutorial is aimed
+at two groups of individuals:
 
 - those familiar with Python (programmers) who want to start to learn game theory,
 -  mathematicians with little or no programming knowledge as a pathway to programming through the interesting
