### 2020-04-25 26a4ba4
adds sentence about a story
Closes #27
