### 2020-04-27 44b079e
Jose review r2 (#29)
* minor copy edits

* adds explanations for the output of notebooks 2 and 3.

Closes #28

* adds sentence about a story

Closes #27
