### 2020-04-27 #78 @Nikoleta-v3
https://github.com/Nikoleta-v3/Game-Theory-and-Python/pull/29 addresses the latest comments @sconde.
