### 2020-05-04 Nikoleta-v3/Game-Theory-and-Python#30 @sconde
JOSS: fix typos
Resolve the typos in the tutorial materials and paper.
