### 2020-05-04 #78 @sconde
https://github.com/Nikoleta-v3/Game-Theory-and-Python/issues/30
