### 2020-05-04 #78 @sconde
@racheltho once the issues have been resolved, and my checklist has been completed, anything else required of me?
