### 2020-05-05 4513263
Fix typos in paper.md
