### 2020-05-05 Nikoleta-v3/Game-Theory-and-Python#30 @drvinceknight
@sconde I've gone through the paper and notebooks and believe I've caught all the typos/errors in #31. I realise the PR diff on the notebook files is not ideal but if you have any further specific ones I've missed and could point them out it would be hugely appreciated.
