### 2020-05-05 2cf9f60
Update nbval to 0.9.5
This is to let us use tags for the nbval-ignore-output which is slightly
cleaner.
