### 2020-05-11 Nikoleta-v3/Game-Theory-and-Python#30 @sconde
Thank you so much
