### 2020-05-11 Nikoleta-v3/Game-Theory-and-Python#30 @drvinceknight
Thank you for your review :+1:
