### 2020-05-17 #78 @racheltho
Hi @mnarayan, You signed up to review for JOSE. Would you like to contribute a review for this submission? The title is "Game Theory and Python: An educational tutorial to game theory and repeated games using Python". Please let me know if you're available and willing. Thanks!
