### 2020-07-20 #78 @Nikoleta-v3
Hello @racheltho 👋 I was wondering when the second reviewer will be assigned to our submission.
