### 2020-11-24 #78 @labarba
hi @drvinceknight — this JOSE submission has suffered from a runaway reviewer. Could you help us out providing a second review? Perhaps you don't have to go so deep, as the first reviewer already completed the review and approved the submission. We just need a second pair of eyes!
Title: **Game Theory and Python: An educational tutorial to game theory and repeated games using Python**
Repo: https://github.com/Nikoleta-v3/Game-Theory-and-Python
