### 2020-11-24 #78 @drvinceknight
Hi @labarba

Here are some suggestions (with context below):

1. `marcharper`
2. `daffidwilde`
3. `fil`
4. `matbesancon`

All of the above have knowledge of game theory and python.

**Disclaimer** There is a potential conflict of interest with the first 2:

- `marcharper` is a core maintainer of the Axelrod library (the library that is used in the educational resource described in the paper under review) as are @Nikoleta-v3 and I. He is also a collaborator of mine on a number of research projects. I'd suggest that his expertise in game theory in general and the Axelrod library in particular (as well his background in pedagogy) makes him an ideal reviewer and knowing him personally I trust that his review would overcome any biases that might exist because of our relationship.
- `daffidwilde` is coming to the end of his PhD under my supervision. However, if you're only needing a "second pair of eyes" I'd again trust his impartial review.

I hope that helps.
