### 2020-11-24 #78 @labarba
LOL so I'm so overwhelmed and dense that i'm asking you to review your own paper. Obviously I didn't click through to the PDF and vaguely remembered that you worked on game theory. Why, yes, you do. 🤷

Thank you for the kindness in your reply, which avoided making me look like a dummy. Well done!
