### 2021-02-19 #78 @labarba
@matbesancon 👋  — You had signed up to be a reviewer for JoSE a while back. Could you take a look at this submission and see if you can contribute a review? (We've been stuck for a number of months with volunteers deluged, so things slowed to a halt. I would really appreciate your help to take this to the finish line.)
**Game Theory and Python: An educational tutorial to game theory and repeated games using Python**
https://github.com/openjournals/jose-reviews/issues/78
