### 2021-03-10 Nikoleta-v3/Game-Theory-and-Python#34 @matbesancon
Notebook 2: add precision that payoff matrix is identical to notebook 1
I was reading the 2nd notebook on matches and tournaments and found myself re-scrolling to be sure I didn't miss the rules of individual games (the payoff matrix), I had to look up the linked doc which indicates that the payoff matrix is indeed the same as in the prisoner's dilemma. Maybe add a precision at the beginning of the notebook that the payoff matrix is identical?

Issue created within the JOSE review https://github.com/openjournals/jose-reviews/issues/78
