### 2021-03-10 #78 @matbesancon
some remarks on the paper itself, this is mostly nitpicking and spelling:

I think the usual spelling is "open-source" as an adjective with a dash
"used in various others events" --> other events
"with a knowledge of Python" --> with knowledge of
"The new strategies then can" -> strategies can then
real life decision making --> real-life decision making
sub fields -> subfields

Would there be a more recent textbook / review article for iterated games to point out in the paper?
