### 2021-03-10 #78 @matbesancon
References: some have the DOI shortcode, some have the DOI URL, this could be made uniform.
The reference Weir 2009 seems oddly formatted (all in caps with a physical address, of the editor? A ... at the end)
