+++ paper.md
@@ -40,14 +40,14 @@ at two groups of individuals:
 subject that is game theory
 
 Game theory is a field of applied mathematics interested in strategic
-interactions. Game theory itself has a number of sub fields and the one
+interactions. Game theory itself has a number of subfields and the one
 considered in this tutorial is the area of repeated games. Repeated games have
 been the subject of research [@Axelrod1981] but have also been used extensively
 as an entry to the subject for students at undergraduate level courses
 [@Brokaw2004], [@Knight2015]. This tutorial focuses on the Iterated Prisoner's
 Dilemma. The advantage of studying the Iterated Prisoner's Dilemma is that it
 models situations in which self-interest clashes with collective interest, thus
-it provides a framework for illustrating the usage of mathematics in real life
+it provides a framework for illustrating the usage of mathematics in real-life
 decision making.
 
 # Content
@@ -67,12 +67,12 @@ but to also serve as an educational and research tool.
 Following the tutorial, participants will be able to reproduce Axelrod's
 tournament but also create their own unique tournaments from a selection of more
 than 200 strategies. Furthermore, participants will be able to progress their
-comprehension of the topic by producing strategies and implement them using
-Python code. The new strategies then can be placed in different tournaments in
+comprehension of the topic by producing strategies and implement them in
+Python. The new strategies can then be placed in different tournaments in
 order to access their success and weakness. This progression of the tutorial
 fits within a constructive framework of learning [@Weir2009] and, by the nature
 of participants actively writing the required software, an active learning
-setting [@Freeman2014]. Participants with a knowledge of Python will benefit from
+setting [@Freeman2014]. Participants with knowledge of Python will benefit from
 an understanding of game theory and participants with no knowledge of Python
 will also be introduced to basic programming concepts through the tangible ideas
 of the Prisoner's Dilemma.
@@ -127,7 +127,7 @@ These instructions can also be found in the [README](https://github.com/Nikoleta
 
 This tutorial was originally formulated to run a game theory workshop at [PyCon
 Namibia 2017](https://na.pycon.org/pycon-namibia-2017/). It has attracted
-the attention of the community ever since and has been used in various others
+the attention of the community ever since and has been used in other
 events.
 
 # Acknowledgements
