### 2021-03-10 ea1f29b
correct and fix style of current references
