+++ paper.bib
@@ -10,11 +10,11 @@
   doi={10.1126/science.7466396},
 }
 
-@misc{Weir2009,
-  title={Approaches to learning: a guide for teachers},
-  author={Weir, Jennifer},
-  year={2009},
-  publisher={ROUTLEDGE JOURNALS, TAYLOR \& FRANCIS LTD 4 PARK SQUARE, MILTON PARK~…},
+@book{Jordan2008,
+  title={Approaches to learning: a guide for teachers: a guide for educators},
+  author={Jordan, Anne and Carlile, Orison and Stack, Annetta},
+  year={2008},
+  publisher={McGraw-Hill Education (UK)}
 }
 
 @article{Freeman2014,
@@ -26,7 +26,7 @@
   pages={8410--8415},
   year={2014},
   publisher={National Acad Sciences},
-  doi={https://doi.org/10.1073/pnas.1319030111}
+  doi={10.1073/pnas.1319030111}
 }
 
 @misc{axelrodproject,
@@ -35,7 +35,6 @@
   month        = apr,
   year         = 2016,
   doi          = {10.5281/zenodo.1405868},
-  url          = {http://dx.doi.org/10.5281/zenodo.<DOI NUMBER>}
 }
 
 @article{Brokaw2004,
@@ -47,7 +46,7 @@
   pages={259--268},
   year={2004},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.3200/JECE.35.3.259-268}
+  doi={10.3200/JECE.35.3.259-268}
 }
 
 @article{Knight2015,
@@ -58,5 +57,5 @@
   number={1},
   pages={28--38},
   year={2015},
-  doi={https://doi.org/10.21100/msor.v14i1.254}
+  doi={10.21100/msor.v14i1.254}
 }
