### 2021-03-10 b162c62
Closes #34
Add information regarding game.
