+++ environment.yml
@@ -8,5 +8,5 @@ dependencies:
     - ipython==6.4.0
     - jupyter==1.0.0
     - docopt
-    - nbval==0.9.5
+    - nbval==0.9.6
     - fsspec>=0.4.3
