+++ requirements.txt
@@ -1,0 +1,7 @@
+axelrod==4.3.0
+ipython==6.4.0
+jupyter==1.0.0
+docopt
+pytest==6.2.2
+nbval==0.9.5
+fsspec>=0.4.3
