### 2021-03-11 37a7b7f
run only on python 3.6 and 3.7
