### 2021-03-11 fddbc80
setup github actions on ubuntu & mac
