+++ paper.bib
@@ -10,6 +10,30 @@
   doi={10.1126/science.7466396},
 }
 
+@article{Nowak2006,
+  title={Five rules for the evolution of cooperation},
+  author={Nowak, Martin A},
+  journal={science},
+  volume={314},
+  number={5805},
+  pages={1560--1563},
+  year={2006},
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.1133755}
+}
+
+@article{Glynatsi2021,
+  title={A bibliometric study of research topics, collaboration, and centrality in the iterated prisoner’s dilemma},
+  author={Glynatsi, Nikoleta E and Knight, Vincent A},
+  journal={Humanities and Social Sciences Communications},
+  volume={8},
+  number={1},
+  pages={1--12},
+  year={2021},
+  publisher={Palgrave},
+  doi={10.1057/s41599-021-00718-9}
+}
+
 @book{Jordan2008,
   title={Approaches to learning: a guide for teachers: a guide for educators},
   author={Jordan, Anne and Carlile, Orison and Stack, Annetta},
