### 2021-03-12 abe0683
Addressing comments of reviewer 2 (#35)
* substitute `open source` with `open-source`

* correct typos

* correct and fix style of current references

* Closes #34

Add information regarding game.

* update version of nbval

* add review papers
