+++ paper.md
@@ -21,11 +21,11 @@ bibliography: paper.bib
 
 # Summary
 
-These materials are an open source educational tutorial aimed at introducing
+These materials are an open-source educational tutorial aimed at introducing
 participants to game theory and more specifically to repeated games. The
-tutorial uses the open source library called
+tutorial uses the open-source library called
 `Axelrod-Python` and [Jupyter
-Notebooks](https://jupyter.org) making this an open source, reproducible and
+Notebooks](https://jupyter.org) making this an open-source, reproducible and
 interactive tutorial.
 
 # Statement of Need
@@ -40,14 +40,14 @@ at two groups of individuals:
 subject that is game theory
 
 Game theory is a field of applied mathematics interested in strategic
-interactions. Game theory itself has a number of sub fields and the one
+interactions. Game theory itself has a number of subfields and the one
 considered in this tutorial is the area of repeated games. Repeated games have
-been the subject of research [@Axelrod1981] but have also been used extensively
+been the subject of research [@Axelrod1981; @Nowak2006; @Glynatsi2021] but have also been used extensively
 as an entry to the subject for students at undergraduate level courses
 [@Brokaw2004], [@Knight2015]. This tutorial focuses on the Iterated Prisoner's
 Dilemma. The advantage of studying the Iterated Prisoner's Dilemma is that it
 models situations in which self-interest clashes with collective interest, thus
-it provides a framework for illustrating the usage of mathematics in real life
+it provides a framework for illustrating the usage of mathematics in real-life
 decision making.
 
 # Content
@@ -60,19 +60,19 @@ In 1980, a political scientist called Robert Axelrod ran a computer
 tournament of the Iterated Prisoner’s Dilemma where strategies written in
 computer code would repeatedly choose between self and collective interest. The
 strategies would decide on their next action using the history of previous
-interactions and the winner was decided based on the average score. The open source
+interactions and the winner was decided based on the average score. The open-source
 package `Axelrod-Python` was created in order to reproduce Axelrod's original work
 but to also serve as an educational and research tool.
 
 Following the tutorial, participants will be able to reproduce Axelrod's
 tournament but also create their own unique tournaments from a selection of more
 than 200 strategies. Furthermore, participants will be able to progress their
-comprehension of the topic by producing strategies and implement them using
-Python code. The new strategies then can be placed in different tournaments in
+comprehension of the topic by producing strategies and implement them in
+Python. The new strategies can then be placed in different tournaments in
 order to access their success and weakness. This progression of the tutorial
-fits within a constructive framework of learning [@Weir2009] and, by the nature
+fits within a constructive framework of learning [@Jordan2008] and, by the nature
 of participants actively writing the required software, an active learning
-setting [@Freeman2014]. Participants with a knowledge of Python will benefit from
+setting [@Freeman2014]. Participants with knowledge of Python will benefit from
 an understanding of game theory and participants with no knowledge of Python
 will also be introduced to basic programming concepts through the tangible ideas
 of the Prisoner's Dilemma.
@@ -127,12 +127,12 @@ These instructions can also be found in the [README](https://github.com/Nikoleta
 
 This tutorial was originally formulated to run a game theory workshop at [PyCon
 Namibia 2017](https://na.pycon.org/pycon-namibia-2017/). It has attracted
-the attention of the community ever since and has been used in various others
+the attention of the community ever since and has been used in other
 events.
 
 # Acknowledgements
 
-The tutorial has been made possible due to the existence of the open source package
+The tutorial has been made possible due to the existence of the open-source package
 [Axelrod-Python](https://github.com/Axelrod-Python/Axelrod). We would like to
 express our appreciation to the maintainers of the package as well as
 anyone that has ever contributed to it.
