+++ paper.bib
@@ -10,11 +10,35 @@
   doi={10.1126/science.7466396},
 }
 
-@misc{Weir2009,
-  title={Approaches to learning: a guide for teachers},
-  author={Weir, Jennifer},
-  year={2009},
-  publisher={ROUTLEDGE JOURNALS, TAYLOR \& FRANCIS LTD 4 PARK SQUARE, MILTON PARK~…},
+@article{Nowak2006,
+  title={Five rules for the evolution of cooperation},
+  author={Nowak, Martin A},
+  journal={science},
+  volume={314},
+  number={5805},
+  pages={1560--1563},
+  year={2006},
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.1133755}
+}
+
+@article{Glynatsi2021,
+  title={A bibliometric study of research topics, collaboration, and centrality in the iterated prisoner’s dilemma},
+  author={Glynatsi, Nikoleta E and Knight, Vincent A},
+  journal={Humanities and Social Sciences Communications},
+  volume={8},
+  number={1},
+  pages={1--12},
+  year={2021},
+  publisher={Palgrave},
+  doi={10.1057/s41599-021-00718-9}
+}
+
+@book{Jordan2008,
+  title={Approaches to learning: a guide for teachers: a guide for educators},
+  author={Jordan, Anne and Carlile, Orison and Stack, Annetta},
+  year={2008},
+  publisher={McGraw-Hill Education (UK)}
 }
 
 @article{Freeman2014,
@@ -26,7 +50,7 @@
   pages={8410--8415},
   year={2014},
   publisher={National Acad Sciences},
-  doi={https://doi.org/10.1073/pnas.1319030111}
+  doi={10.1073/pnas.1319030111}
 }
 
 @misc{axelrodproject,
@@ -35,7 +59,6 @@
   month        = apr,
   year         = 2016,
   doi          = {10.5281/zenodo.1405868},
-  url          = {http://dx.doi.org/10.5281/zenodo.<DOI NUMBER>}
 }
 
 @article{Brokaw2004,
@@ -47,7 +70,7 @@
   pages={259--268},
   year={2004},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.3200/JECE.35.3.259-268}
+  doi={10.3200/JECE.35.3.259-268}
 }
 
 @article{Knight2015,
@@ -58,5 +81,5 @@
   number={1},
   pages={28--38},
   year={2015},
-  doi={https://doi.org/10.21100/msor.v14i1.254}
+  doi={10.21100/msor.v14i1.254}
 }
