### 2021-03-12 1e27df3
GitHub actions (#36)
* remove travis.yml

* set up github actions

* update nbval version

* run on push

* run only on python 3.6 and 3.7

* setup github actions on ubuntu & mac
