+++ .travis.yml
@@ -1,27 +1,0 @@
-language: python
-python:
-  - "3.6"
-before_install:
-  - export MPLBACKEND="agg"
-dist: xenial
-services:
-  - xvfb
-install:
-  - sudo apt-get update
-  # We do this conditionally because it saves us some downloading if the
-  # version is the same.
-  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH="$HOME/miniconda/bin:$PATH"
-  - hash -r
-  - conda config --set always_yes yes --set changeps1 no
-  - conda update -q conda
-  # Useful for debugging any issues with conda
-  - conda info -a
-
-  - conda env create -f environment.yml
-  - source activate game-python
-
-script:
-
-  - sh run_tests.sh
