### 2021-03-12 #78 @Nikoleta-v3
Hello @matbesancon I believe I have now addressed all the issues that you raised 😄

This was done in https://github.com/Nikoleta-v3/Game-Theory-and-Python/pull/35.

Note I've changed the CI to GitHub actions (https://github.com/Nikoleta-v3/Game-Theory-and-Python/pull/36) 😬
