### 2021-03-15 82f915e
run tests only on pull requests
