+++ .github/workflows/ci.yml
@@ -1,6 +1,6 @@
 name: ci
 
-on: [push]
+on: [pull_request]
 
 jobs:
   build:
