### 2021-05-23 #78 @labarba
Hi everybody!
Thank you to the Reviewers: @sconde , @matbesancon 🙏

@Nikoleta-v3 — Could you now up the version, issue a tagged release in your repo, and make a Zenodo deposit? Report the version number here, as well as the Zenodo DOI, so we can run the commands to update that metadata for the submission.
