### 2021-05-23 #78 @labarba
I have a question: can this workshop _really_ be completed in two hours? It seems that a thoughtful engagement with all of the material and exercises would take longer, maybe even double that amount of time!
