### 2021-05-24 df28a57
adds statement regarding the duration of the workshop
