### 2021-05-24 #78 @Nikoleta-v3
Hello @labarba. Thank you for getting back to us.

Regarding your question, we recommend 2 hours because we have completed the workshop in 2 hours at a PyCon where the audience was proficient with Python. However, we realise that might not be the case each time.  We have reworded the text to say:

>>>  Though we recommend that the material can be covered in 2 hours, in both a workshop environment and through independent learning, some audience/learners might not complete all the exercises. We recommend any facilitator/learner to adjust as necessary.

Please let us know what you think. Once you have given me the okay I will go ahead and create a new release and make a Zenodo deposit.
