### 2021-05-24 #78 @labarba
> some audience/learners might ~~not complete~~ need more time to complete all the exercises
