### 2021-05-24 #78 @Nikoleta-v3
Thank you for the great suggestion. I have address it here: https://github.com/Nikoleta-v3/Game-Theory-and-Python/pull/40/files
