### 2021-05-24 #78 @Nikoleta-v3
The version number is **v1.0.0**: https://github.com/Nikoleta-v3/Game-Theory-and-Python/releases/tag/v1.0.0)

The Zenodo DOI is **10.5281/zenodo.4783322**: https://zenodo.org/record/4783322#.YKum0y8RpTY
